Adding onto SEIR Model
Now have a different classes for Asymptomatic vs. symptomatic infecteds. Still have the other 4 classes of S, E, R, D. The current adjutsable values for the user are both of the contact rates for asymptomatic and symptomatic individuals with susceptibles, the rate of recovery, and death rate for symptomatic infecteds. The current model is based off of a population of 1,000 people and the time unit is days