Updates and Additions to Working_SEIR.py
Added a second graph that just shows the 4 infected classes. Made testing rate a function of time and user can adjust the rate at which testing is done. Added a table with parameter names and their current values. Added a tab that has a description of how the model works. Changed y-axis from proportion of population to # of individuals