Updating network nodes and continuous equations for line graphs
In Working_SEIR.py I changed the piecewise functions for the vaccine frequency and the health capacity to be contoniuous sigmoid functions rather than being piecewise functions for better integration.

In working_bubble.py the slider for time now works and updates the sizes of the nodes for each class as time progresses. The "Play" button dosen't work properly yet