Animation and Combining Files
The "Play" button in working_bubble.py now works and shows the node sizes changing over time.

There are three different python files for the three different tabs that I currently have for this model. All of these files are combines and the tabs are organized in the SIR.py file

SIR.py is the file to be run in order to see the output for this model