Layout and Bar Graph
Gave graphs sizing_mode values of "scale_both" that should hopefully help with the fit of models on the webpage.

Added a bar graph next to the network graph that shows proportions of the population in each class. This animates as the same time as the network graph but it might be easier for people to process the sizing of the bars rather than the circle sizes