Using dataframe in ZECC
I converted almost all of the ZECC file to use pandas dataframes instead. Still havng an issue with it reading lists in as strings and then not being able to store lists as individual values in the data frame.
Also added .csv file to folder for access to location weather info