Finishing transfer over to Pandas dataframe
Now all location information for ZECC is stored in csv files that are then converted to pandas dataframes to use in the code. All of ZECC.py has been adjusted for the functionality to work with a dataframe set up.

The notebook for ZECC has been updated to reflect ZECC.py